[MOBI-33-landing-page-bottom-mobile-publishing] Refactor: 상수 파일명 케밥 케이스로 수정, 함수 landing util로 이동
