+++ apps/academy/src/widget/landing/constants/curriculum-schedule.ts
