+++ apps/academy/src/widget/landing/constants/stack-images.ts
