+++ apps/academy/src/widget/landing/util/index.ts
@@ -1,0 +1,1 @@
+export { getBorderRadius } from './get-border-radius'
