+++ packages/util/src/index.ts
@@ -1,4 +1,3 @@
 export { cn } from './cn'
 export { blankOpener } from './blank-opener'
 export { formatDate } from './date-formatter'
-export { getBorderRadius } from './get-border-radius'
